requirements.txt
alembic==1.16.1
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.3
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gt==0.3.15
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dotenv==1.1.0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
QtPy==2.4.3
requests==2.32.4
setuptools==80.9.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3